Remove MiscAPI.java
This was replaced by MiscUtil.java
